Fix refactoring issues
